requirements.txt
pillow
python-dotenv
pytz
requests
apscheduler
